Update mod loading structure.
